% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp11tesseract-package.R
\docType{package}
\name{cpp11tesseract-package}
\alias{cpp11tesseract}
\alias{cpp11tesseract-package}
\title{Open Source OCR Engine}
\description{
Bindings to 'Tesseract':
a powerful optical character recognition (OCR) engine that supports over 100
languages. The engine is highly configurable in order to tune the detection
algorithms and obtain the best possible results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pacha.dev/cpp11tesseract/}
  \item Report bugs at \url{https://github.com/pachadotdev/cpp11tesseract/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.sepulveda@mail.utoronto.ca} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

Authors:
\itemize{
  \item Jeroen Ooms \email{jeroen@berkeley.edu} (\href{https://orcid.org/0000-0002-4035-0289}{ORCID})
}

Other contributors:
\itemize{
  \item Munk School of Global Affairs and Public Policy [funder]
}

}
