% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tesseract.R
\name{pdf_to_png}
\alias{pdf_to_png}
\title{Export a PDF file to PNG files}
\usage{
pdf_to_png(path, dpi = 600)
}
\arguments{
\item{path}{path to the PDF file}

\item{dpi}{resolution in DPI}
}
\value{
a "magick-image" object
}
\description{
Export a PDF file to PNG files
}
\examples{
if (requireNamespace("magick", quietly = TRUE)) {
 file <- system.file("examples", "ocrscan.pdf", package = "cpp11tesseract")
 pdf_to_png(file)
}
}
